requirements.txt
fastapi
uvicorn[standard]
pillow
torch
transformers
huggingface_hub
python-multipart